+++ config/application.rb
@@ -1,22 +1,23 @@
-require_relative 'boot'
+# frozen_string_literal: true
 
+require_relative 'boot'
 require 'rails/all'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
 
-module QaServer
+module BclUpServer
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.1
+    config.load_defaults 7.0
     config.web_console.development_only = false if config.respond_to?(:web_console) && !Rails.env.production?
 
     # Qa::Authorities::Discogs::GenericAuthority.discogs_key = ENV['DISCOGS_KEY']
     # Qa::Authorities::Discogs::GenericAuthority.discogs_secret = ENV['DISCOGS_SECRET']
 
     config.to_prepare do
-      QaServer::AuthorityListPresenter.prepend PrependedPresenters::AuthorityListPresenter
+      BclUpServer::AuthorityListPresenter.prepend PrependedPresenters::AuthorityListPresenter
       Qa::Authorities::LinkedData::FindTerm.prepend PrependedLinkedData::FindTerm
       Qa::Authorities::LinkedData::SearchQuery.prepend PrependedLinkedData::SearchQuery
       RDF::Graph.prepend PrependedRdf::RdfGraph