+++ CHANGELOG.md
@@ -12,4 +12,4 @@ Release v1.0.2 (2022-04-15)
 * Release v1.0.1 (2022-04-15)
 
 * add GitHub actions to auto-build and deploy images to ECR
-* Release of base qa_server as a Docker container image.
+* Release of base bcl_up_server as a Docker container image.